# Conflicts:
#	bookstore/src/main/java/pq/jdev/b001/bookstore/Category/repository/CategoryRepository.java
#	bookstore/src/main/java/pq/jdev/b001/bookstore/Category/service/CategoryAddEditServiceImpl.java
#	bookstore/src/main/resources/templates/categoryList.html

+++ bookstore/src/main/java/pq/jdev/b001/bookstore/Category/model/Category.java
@@ -30,7 +30,7 @@ import pq.jdev.b001.bookstore.books.model.Book;
 public class Category {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private long id;
+	private int id;
 	private String name;
 	private Date createdate;
 	private Date updatedate;
@@ -39,7 +39,7 @@ public class Category {
 		
 	}
 	
-	public Category(long id, String name, Date createdate,Date updatedate) {
+	public Category(int id, String name, Date createdate,Date updatedate) {
 		this.id = id;
 		this.name = name;
 		this.createdate = createdate;
@@ -49,10 +49,10 @@ public class Category {
 	@ManyToMany(cascade = {CascadeType.PERSIST, CascadeType.MERGE}, mappedBy = "categories", targetEntity = pq.jdev.b001.bookstore.books.model.Book.class)
 	private Set<Book> books;
 	
-	public long getId() {
+	public int getId() {
 		return id;
 	}
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 	public String getName() {
